Downloads/Project-Chimera/requirements.txt
# Project Chimera - Consolidated Requirements
# Advanced AI-Driven Mashup Creation Platform

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# AI Collaboration Engines  
openai>=1.3.0
anthropic>=0.20.0

# Core Audio Processing
librosa>=0.10.1
numpy>=1.24.0
scipy>=1.10.0
pydub>=0.25.1
soundfile>=0.12.1

# Professional Audio Engines
pyrubberband>=0.3.0     # High-quality time-stretching/pitch-shifting
pedalboard>=0.8.0       # Professional effects chains

# AI Source Separation (Multiple Models)
spleeter>=2.3.0         # Google's separation model
demucs>=4.0.0           # Facebook's state-of-the-art separation

# Audio Acquisition & Search  
yt-dlp>=2023.7.6        # YouTube audio download
thefuzz>=0.19.0         # Fuzzy string matching for search
python-Levenshtein>=0.20.9  # String distance for search optimization

# Data Processing & Utilities
scikit-learn>=1.3.0     # Machine learning utilities
matplotlib>=3.7.0       # Visualization (for analysis)
tqdm>=4.65.0            # Progress bars
pathlib2>=2.3.7         # Path handling
joblib>=1.3.0           # Parallel processing

# Optional: Advanced ML Models
# torch>=2.0.0          # PyTorch for neural models (uncomment if using torch-based models)
# torchaudio>=2.0.0     # PyTorch audio processing